requirements.txt
matplotlib~=3.5.1
numpy~=1.21.5
Pillow~=9.0.0
